WIP refactor js articles files
