+++ openaev-front/src/components/fields/AutocompleteField.tsx
@@ -185,6 +185,7 @@ const AutocompleteField: FunctionComponent<Props> = (props) => {
           if (custom === null) {
             return null;
           }
+          return custom;
         }
 
         return defaultRenderOption(liProps, option);
